Use `Path` class

+++ package/solie/utility/examine_data_files.py
@@ -1,5 +1,6 @@
 import json
 import os
+from pathlib import Path
 
 import aiofiles
 import pandas as pd
@@ -9,16 +10,19 @@ from solie.utility import user_settings
 
 
 async def do():
-    datapath = user_settings.get_app_settings()["datapath"]
+    datapath_str = user_settings.get_app_settings()["datapath"]
+    if datapath_str is None:
+        return
+    datapath = Path(datapath_str)
 
     # 5.0: data settings
     try:
-        filepath = f"{datapath}/basics.json"
+        filepath = datapath / "basics.json"
         async with aiofiles.open(filepath, "r", encoding="utf8") as file:
             content = await file.read()
             data_settings = json.loads(content)
         os.remove(filepath)
-        filepath = f"{datapath}/data_settings.json"
+        filepath = datapath / "data_settings.json"
         async with aiofiles.open(filepath, "w", encoding="utf8") as file:
             content = json.dumps(data_settings, indent=4)
             await file.write(content)
@@ -27,7 +31,7 @@ async def do():
 
     # 5.0: symbol column was added to auto order record
     try:
-        filepath = f"{datapath}/transactor/auto_order_record.pickle"
+        filepath = datapath / "transactor" / "auto_order_record.pickle"
         auto_order_record: pd.DataFrame = await go(pd.read_pickle, filepath)
         if "Symbol" not in auto_order_record.columns:
             auto_order_record["Symbol"] = ""
@@ -37,7 +41,7 @@ async def do():
 
     # 5.0: solie default strategy now has strategy code SLSLDS
     try:
-        filepath = f"{datapath}/transactor/automation_settings.json"
+        filepath = datapath / "transactor" / "automation_settings.json"
         async with aiofiles.open(filepath, "r", encoding="utf8") as file:
             content = await file.read()
             automation_settings = json.loads(content)
@@ -52,7 +56,7 @@ async def do():
 
     # 6.0: use strategy index instead of strategy code
     try:
-        filepath = f"{datapath}/transactor/automation_settings.json"
+        filepath = datapath / "transactor" / "automation_settings.json"
         async with aiofiles.open(filepath, "r", encoding="utf8") as file:
             content = await file.read()
             automation_settings = json.loads(content)
@@ -68,7 +72,7 @@ async def do():
 
     # 6.3: auto_trade and manual_trade
     try:
-        filepath = f"{datapath}/transactor/asset_record.pickle"
+        filepath = datapath / "transactor" / "asset_record.pickle"
         asset_record: pd.DataFrame = await go(pd.read_pickle, filepath)
         asset_record["Cause"] = asset_record["Cause"].replace("trade", "auto_trade")
         await go(asset_record.to_pickle, filepath)
